test(e2e): run the watch page tests offline
Thirteen more watch tests only needed the live API because their fixtures sat
in another test's directory: recordings are keyed per test title, so the page
and comment responses for "Me at the zoo" were invisible to every test but the
three that recorded them. Pooling those directories in the mocked watch page
lets the whole set run offline, against recorded metadata and the demo video.

They keep their assertions; only the setup changes. Each registers the page
mock before its own routes, since the mock would otherwise shadow them - two
of them were previously overriding a response that never reached the app.

Four stay in the network suite because they need what the demo video cannot
provide (storyboards for the progress preview, real segment timing), and two
more because they raced at offline speed. Skipped network tests are down from
50 to 37.
